rename file as well
